requirements.txt
# requirements.txt
# Core deep learning libraries
torch>=1.9.0
torchvision>=0.10.0

# Image processing and augmentation
opencv-python>=4.5.0
albumentations>=1.0.0
Pillow>=8.0.0

# Data handling
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Evaluation metrics
pycocotools>=2.0.4
scikit-learn>=0.24.0

# Progress bars and UI
tqdm>=4.62.0
rich>=10.0.0

# Utilities
glob2>=0.7
lxml>=4.6.0

# Optional: For better performance
# pillow-simd  # Faster image processing (install manually if needed)